pyproject.toml
# Python Project Configuration
# This file configures the build system, dependencies, and development tools
# for the Material Diffusion Blender add-on project

[project]
name = "blender-material-diffusion"
version = "1.0.0"
description = "AI-powered material generation add-on for Blender using ComfyUI backend"
readme = "README.md"
authors = [
    {name = "Eugenio Buffo", email = "buffo.eugenio@gmail.com"}
]
maintainers = [
    {name = "Eugenio Buffo", email = "buffo.eugenio@gmail.com"}
]
license = {file = "LICENSE"}
keywords = [
    "blender", 
    "addon", 
    "extension",
    "texture", 
    "material",
    "diffusion", 
    "comfyui", 
    "ai", 
    "generative",
    "3d",
    "rendering"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Other Environment",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Topic :: Multimedia :: Graphics :: 3D Rendering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
    "requests>=2.31.0"
]

[project.urls]
"Homepage" = "https://github.com/EugenioBuffo/blender-material-diffusion"
"Repository" = "https://github.com/EugenioBuffo/blender-material-diffusion"
"Bug Tracker" = "https://github.com/EugenioBuffo/blender-material-diffusion/issues"
"Documentation" = "https://github.com/EugenioBuffo/blender-material-diffusion#readme"
"Changelog" = "https://github.com/EugenioBuffo/blender-material-diffusion/releases"

[project.optional-dependencies]
dev = [
    # Blender development dependencies
    "fake-bpy-module-latest>=20240604",
    
    # Core dependencies for development
    "Pillow>=10.0.0",
    "requests>=2.31.0",

]

# Build System Configuration
[build-system]
requires = [
    "setuptools>=68.0.0",
    "wheel>=0.41.0"
]
build-backend = "setuptools.build_meta"

# Tool Configurations
[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip_gitignore = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
